refactor: shorten import code style minors
